refactor: move function validateDate to useAppointmentProvider
