+++ src/components/FormRegister/index.js
@@ -14,7 +14,7 @@ import { useFormik, Formik, useFormikContext, useField } from 'formik';
 
 
 const FormRegister = () => {
-    const { newAppointmentData, setNewAppointmentData, alertMessage, toggleAlertMessage, registerAppointment, alertStatus } = useAppointmentProvider();
+    const { newAppointmentData, setNewAppointmentData, alertMessage, toggleAlertMessage, registerAppointment, alertStatus, validateDate } = useAppointmentProvider();
 
     const [error, setError] = useState({
         name: false,
@@ -69,18 +69,6 @@ const FormRegister = () => {
     const isWeekday = (date) => {
         const day = getDay(date);
         return day !== 0 && day !== 6;
-    }
-
-    const validateDate = (dateValue) => {
-        const newDate = [];
-        if(typeof(dateValue) === 'string') {
-            return dateValue;
-        }
-
-        getDate(dateValue).toString().length === 1 ? newDate.push(`0${getDate(dateValue)}`) : newDate.push(getDate(dateValue).toString());
-        getMonth(dateValue).toString().length === 1 ? newDate.push(`0${getMonth(dateValue)+1}`) : newDate.push((getMonth(dateValue)+1).toString());
-        newDate.push(getYear(dateValue).toString());
-        return newDate.join('/');
     }
 
     return (
